+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Add support for printing to the current buffer to `cider-eval-defun-up-to-point`.
 * Remove `cider-ping` command.
 * Remove `cider-visit-error-buffer` in favour of using `cider-selector`.
+* Rename `cider-refresh` to `cider-ns-refresh` (and all the related defcustoms).
 
 ## 0.17.0 (2018-05-07)